feat: circuit breaker
