+++ src/main/kotlin/no/digdir/fdk/searchservice/Application.kt
@@ -2,9 +2,11 @@ package no.digdir.fdk.searchservice
 
 import org.springframework.boot.SpringApplication
 import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.autoconfigure.security.servlet.UserDetailsServiceAutoConfiguration
 import org.springframework.boot.context.properties.ConfigurationPropertiesScan
 
-@SpringBootApplication
+
+@SpringBootApplication(exclude = [UserDetailsServiceAutoConfiguration::class])
 @ConfigurationPropertiesScan
 open class Application
