+++ src/main/kotlin/no/digdir/fdk/searchservice/kafka/KafkaRdfParseEventConsumer.kt
@@ -1,86 +1,35 @@
 package no.digdir.fdk.searchservice.kafka
 
-import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
-import io.micrometer.core.instrument.Metrics
 import no.digdir.fdk.searchservice.elastic.SearchRepository
-import no.digdir.fdk.searchservice.mapper.toSearchObject
-import no.digdir.fdk.searchservice.model.*
 import no.fdk.rdf.parse.RdfParseEvent
-import no.fdk.rdf.parse.RdfParseResourceType
 import org.apache.kafka.clients.consumer.ConsumerRecord
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import org.springframework.kafka.annotation.KafkaListener
 import org.springframework.kafka.support.Acknowledgment
 import org.springframework.stereotype.Component
-import kotlin.time.measureTimedValue
-import kotlin.time.toJavaDuration
+import java.time.Duration
 
 
 @Component
 class KafkaRdfParseEventConsumer(
-    private val searchRepository: SearchRepository
+    private val searchRepository: SearchRepository,
+    private val kafkaRdfParseEventCircuitBreaker: KafkaRdfParseEventCircuitBreaker
 ) {
-    private fun <T> index(event: RdfParseEvent, clazz: Class<T>, index: (T) -> Unit) {
-        val search = searchRepository.findById("${event.fdkId}")
-        if (search.isEmpty || (search.get().metadata?.timestamp ?: 0) < event.timestamp) {
-            val payload = jacksonObjectMapper().readValue(event.data.toString(), clazz)
-            index(payload)
-        }
-    }
 
     @KafkaListener(
         topics = ["rdf-parse-events"],
         groupId = "fdk-search-service",
-        containerFactory = "kafkaListenerContainerFactory"
+        containerFactory = "kafkaListenerContainerFactory",
+        concurrency = "4",
+        id = "rdf-parse"
     )
     fun listen(record: ConsumerRecord<String, RdfParseEvent>, ack: Acknowledgment) {
-        LOGGER.debug("Received message - offset: " + record.offset())
-
-        val event = record.value()
         try {
-            val timeElapsed = measureTimedValue {
-                if (event?.resourceType == RdfParseResourceType.DATASET) {
-                    LOGGER.debug("Index dataset - id: " + event.fdkId)
-                    index(event, Dataset::class.java) {
-                        searchRepository.save(it.toSearchObject("${event.fdkId}", event.timestamp))
-                    }
-                } else if (event?.resourceType == RdfParseResourceType.DATA_SERVICE) {
-                    LOGGER.debug("Index dataservice - id: " + event.fdkId)
-                    index(event, DataService::class.java) {
-                        searchRepository.save(it.toSearchObject("${event.fdkId}", event.timestamp))
-                    }
-                } else if (event?.resourceType == RdfParseResourceType.CONCEPT) {
-                    LOGGER.debug("Index concept - id: " + event.fdkId)
-                    index(event, Concept::class.java) {
-                        searchRepository.save(it.toSearchObject("${event.fdkId}", event.timestamp))
-                    }
-                } else if (event?.resourceType == RdfParseResourceType.INFORMATION_MODEL) {
-                    LOGGER.debug("Index informationmodel - id: " + event.fdkId)
-                    index(event, InformationModel::class.java) {
-                        searchRepository.save(it.toSearchObject("${event.fdkId}", event.timestamp))
-                    }
-                } else if (event?.resourceType == RdfParseResourceType.EVENT) {
-                    LOGGER.debug("Index event - id: " + event.fdkId)
-                    index(event, Event::class.java) {
-                        searchRepository.save(it.toSearchObject("${event.fdkId}", event.timestamp))
-                    }
-                } else if (event?.resourceType == RdfParseResourceType.SERVICE) {
-                    LOGGER.debug("Index service - id: " + event.fdkId)
-                    index(event, Service::class.java) {
-                        searchRepository.save(it.toSearchObject("${event.fdkId}", event.timestamp))
-                    }
-                }
-                ack.acknowledge()
-            }
-            Metrics.timer("search_index", "type", event.resourceType.name.lowercase())
-                .record(timeElapsed.duration.toJavaDuration())
+            kafkaRdfParseEventCircuitBreaker.process(record, searchRepository)
+            ack.acknowledge()
         } catch (e: Exception) {
-            LOGGER.error("Error processing message: " + e.message)
-            Metrics.counter(
-                "search_index_error",
-                "type", event.resourceType.name.lowercase()
-            ).increment()
+            ack.nack(Duration.ZERO)
         }
     }
